tradehub-fund/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
amqp==5.3.1
    # via kombu
asgiref==3.11.1
    # via
    #   django
    #   django-cors-headers
billiard==4.2.4
    # via celery
celery==5.6.2
    # via
    #   fundval (pyproject.toml)
    #   django-celery-beat
    #   django-celery-results
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
chinese-calendar==1.11.0
    # via fundval (pyproject.toml)
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
django==6.0.2
    # via
    #   fundval (pyproject.toml)
    #   django-celery-beat
    #   django-celery-results
    #   django-cors-headers
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
django-celery-beat==2.1.0
    # via fundval (pyproject.toml)
django-celery-results==2.6.0
    # via fundval (pyproject.toml)
django-cors-headers==4.9.0
    # via fundval (pyproject.toml)
django-timezone-field==4.2.3
    # via django-celery-beat
djangorestframework==3.16.1
    # via
    #   fundval (pyproject.toml)
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via fundval (pyproject.toml)
gunicorn==25.0.3
    # via fundval (pyproject.toml)
idna==3.11
    # via requests
kombu==5.6.2
    # via celery
packaging==26.0
    # via
    #   gunicorn
    #   kombu
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via fundval (pyproject.toml)
pyjwt==2.11.0
    # via djangorestframework-simplejwt
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.2.1
    # via fundval (pyproject.toml)
pytz==2025.2
    # via django-timezone-field
redis==7.1.1
    # via fundval (pyproject.toml)
requests==2.32.5
    # via fundval (pyproject.toml)
six==1.17.0
    # via python-dateutil
sqlparse==0.5.5
    # via django
tzdata==2025.3
    # via kombu
tzlocal==5.3.1
    # via celery
urllib3==2.6.3
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.6.0
    # via prompt-toolkit
